GROUP B-Step1: main urls.py and languages/urls.py
Edit the main urls.py file by adding 'include' and putting it between import and path so I could include languages app urls. I add an index path 'path('', include('languages.urls))' to the file, so when people go to the site they first go to the main url and then go to the languages url if they asked for the languages.urls.
Create a new file under languages app and call it 'urls.py'. In the languages/urls.py file add the import views module and the routers from rest_framework. The routers is going to take care of generating all the urls for my particular model. When i go to my url i type urls.com/languages if I perform a get request I expect to see all of the different languages in my database, if I perform a post request on that same url then I expect to create a new language, if i do 'languages/id' or 'languages/1' I asking for the first language in the database so instead of doing all of that django rest framework makes it easier for us to generate this if we are doing it in the very standard consistent way, so to do this I am going to create a router and the first argument is the endpoint with is 'languages' and the second argument is calling the url to send the request to which is views.LanguageView. Now the path in the array is index and what it includes is router.urls