GROUP E-Step1: Serializer
What a serializer does in the context of an API it translates to and from JSON or any format you can transfer over http. A JSON object is contained of list and dictionaries, it is arrays and objects but in Python it is list and dictionaries so you send a list over a request and a serializer will deserializer that combination of list and dictionaries into whatever your application object is, in our case a language. You can't send a language model over an http it only exist in your app so you have to send a common format which is JSON so a serializer will translate over to and from JSON.
Create a file called 'serializers.py' and inside the file is import serializers from the rest_framework and then import Language model from .models.
Next is to create a class and call it 'LanguageSerializer' then below that I create a Meta class and fill it with a model which = Languge and fields. These are the attributes. The 3 fields will be displayed.