GROUP F-Step1: views.py
Make 3 imports then create a LanguageView class with a ModelViewSet that inherits from viewsets.  queryset is asking how i am goingn to get all the data from the database so you all of the Language model.The serializer class is going to be the language serializer class in the serializers.py file.